Updated some test scripts.
